完了: 見積回答タブ機能の100%実装
- 見積回答の新規作成、一覧表示、編集機能を実装
- 必須項目（ベンダー、見積金額、納期、有効期限、添付ファイル）をすべて実装
- データベーススキーマの更新とバックエンドAPI統合
- 日付フォーマット問題を解決し、フォーム送信が正常動作
- UIコンポーネント（Badge、Alert、Table）を追加
- 承認済み見積依頼の選択機能を実装
- ファイルアップロード機能を統合

実装内容:
- EstimateResponse モデルの更新（estimate_number, estimate_price等の追加）
- estimate_response_service_impl.rs の SQL クエリ更新
- 新規回答作成ページ（/responses/new）の完全実装
- 回答一覧ページ（/responses）の更新
- 日付処理の修正とフォームバリデーション強化

テスト結果:
- フォーム送信: ✅ 成功
- データベース保存: ✅ 成功
- 一覧表示: ✅ 成功
- 日付フォーマット: ✅ 修正完了
- API統合: ✅ 動作確認済み

+++ src/backend/src/models/estimate_response.rs
@@ -2,6 +2,7 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 use validator::Validate;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct EstimateResponse {
@@ -9,8 +10,8 @@ pub struct EstimateResponse {
     pub request_id: i32,
     pub vendor_id: Option<i32>,
     pub estimate_number: Option<String>,
-    pub estimate_price: Option<f64>,
-    pub total_amount: Option<f64>,
+    pub estimate_price: Option<Decimal>,
+    pub total_amount: Option<Decimal>,
     pub breakdown: Option<serde_json::Value>,
     pub delivery_date: Option<chrono::NaiveDate>,
     pub validity_period: Option<String>,
@@ -29,8 +30,8 @@ pub struct CreateEstimateResponseRequest {
     pub vendor_id: Option<i32>,
     #[validate(length(max = 50))]
     pub estimate_number: Option<String>,
-    pub estimate_price: Option<f64>,
-    pub total_amount: Option<f64>,
+    pub estimate_price: Option<Decimal>,
+    pub total_amount: Option<Decimal>,
     pub breakdown: Option<serde_json::Value>,
     pub delivery_date: Option<String>,
     pub validity_period: Option<String>,
@@ -46,8 +47,8 @@ pub struct UpdateEstimateResponseRequest {
     pub vendor_id: Option<i32>,
     #[validate(length(max = 50))]
     pub estimate_number: Option<String>,
-    pub estimate_price: Option<f64>,
-    pub total_amount: Option<f64>,
+    pub estimate_price: Option<Decimal>,
+    pub total_amount: Option<Decimal>,
     pub breakdown: Option<serde_json::Value>,
     pub delivery_date: Option<String>,
     pub validity_period: Option<String>,
